calculator

+++ python -practise/2nd.py
@@ -1,6 +1,6 @@
 num_1 = input("First Number ")
-number_1=int(num_1);
+number_1=float(num_1);
 num_2 = input("Second Number ")
-number_2=int(num_2);
+number_2=float(num_2);
 sum=number_1+number_2
 print("Sum is : ",sum)
